设置统一`HooksType`，向外解耦所挂钩子
